+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/configuresourcescan_client.go
@@ -19,7 +19,7 @@ import (
 // ConfigureSourceScanClient contains the methods for the ConfigureSourceScan group.
 // Don't use this type directly, use NewConfigureSourceScanClient() instead.
 //
-// Generated from API version 2026-07-01
+// Generated from API version 2026-08-01
 type ConfigureSourceScanClient struct {
 	internal       *arm.Client
 	subscriptionID string
@@ -30,6 +30,9 @@ type ConfigureSourceScanClient struct {
 //   - credential - used to authorize requests. Usually a credential from azidentity.
 //   - options - Contains optional client configuration. Pass nil to accept the default values.
 func NewConfigureSourceScanClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*ConfigureSourceScanClient, error) {
+	if subscriptionID == "" {
+		return nil, errors.New("parameter subscriptionID cannot be empty")
+	}
 	cl, err := arm.NewClient(moduleName, moduleVersion, credential, options)
 	if err != nil {
 		return nil, err
@@ -96,7 +99,7 @@ func (client *ConfigureSourceScanClient) execute(ctx context.Context, resourceGr
 func (client *ConfigureSourceScanClient) executeCreateRequest(ctx context.Context, resourceGroupName string, vaultName string, fabricName string, containerName string, protectedItemName string, body ProtectedItemConfigureSourceScanRequest, _ *ConfigureSourceScanClientBeginExecuteOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.RecoveryServices/vaults/{vaultName}/backupFabrics/{fabricName}/protectionContainers/{containerName}/protectedItems/{protectedItemName}/configureSourceScan"
 	if client.subscriptionID == "" {
-		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+		return nil, errors.New("parameter subscriptionID cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
 	if resourceGroupName == "" {
@@ -124,7 +127,7 @@ func (client *ConfigureSourceScanClient) executeCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", version20260701)
+	reqQP.Set("api-version", version20260801)
 	req.Raw().URL.RawQuery = strings.ReplaceAll(reqQP.Encode(), "+", "%20")
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}