+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/featuresupport_client_example_test.go
@@ -12,7 +12,7 @@ import (
 	"log"
 )
 
-// Generated from example definition: 2026-07-01/AzureIaasVm/BackupFeature_Validate.json
+// Generated from example definition: 2026-08-01/AzureIaasVm/BackupFeature_Validate.json
 func ExampleFeatureSupportClient_Validate() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {