+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/jobdetails_client_example_test.go
@@ -11,7 +11,7 @@ import (
 	"log"
 )
 
-// Generated from example definition: 2026-07-01/Common/GetJobDetails.json
+// Generated from example definition: 2026-08-01/Common/GetJobDetails.json
 func ExampleJobDetailsClient_Get() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {