+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/protecteditemoperationresults_client_example_test.go
@@ -11,7 +11,7 @@ import (
 	"log"
 )
 
-// Generated from example definition: 2026-07-01/AzureIaasVm/ProtectedItemOperationResults.json
+// Generated from example definition: 2026-08-01/AzureIaasVm/ProtectedItemOperationResults.json
 func ExampleProtectedItemOperationResultsClient_Get() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {