+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 5.1.0 (2026-08-29)
+### Features Added
+
+- New function `*ItemLevelRecoveryConnectionsClient.ListInstantItemRecoveryOperationResult(ctx context.Context, resourceGroupName string, vaultName string, fabricName string, containerName string, protectedItemName string, recoveryPointID string, body InstantItemRecoveryOperationResultRequest, options *ItemLevelRecoveryConnectionsClientListInstantItemRecoveryOperationResultOptions) (ItemLevelRecoveryConnectionsClientListInstantItemRecoveryOperationResultResponse, error)`
+- New struct `InstantItemRecoveryOperationResultRequest`
+
+
 ## 5.0.0 (2026-08-13)
 ### Breaking Changes